@@ -13,7 +13,7 @@
         "doctrine/dbal": "^3.1",
         "guzzlehttp/guzzle": "^7",
         "laravel-notification-channels/telegram": "^6.0",
-        "laravel/framework": "^11.0",
+        "laravel/framework": "^12.0",
         "laravel/sanctum": "^4.0",
         "laravel/tinker": "^2",
         "laravel/ui": "^4.0",
@@ -27,16 +27,16 @@
         "vlucas/phpdotenv": "^5.4"
     },
     "require-dev": {
-        "driftingly/rector-laravel": "^0.29.0",
+        "driftingly/rector-laravel": "^2.0",
         "fakerphp/faker": "^1.9.1",
+        "larastan/larastan": "^3.0",
         "laravel/pint": "^1.1",
         "laravel/sail": "^1.0.1",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/larastan": "^2.0",
-        "pestphp/pest": "^2.0",
-        "pestphp/pest-plugin-laravel": "^2.0",
-        "phpunit/phpunit": "^10.1",
-        "rector/rector": "^0.18.0"
+        "pestphp/pest": "^3.8",
+        "pestphp/pest-plugin-laravel": "^3.2",
+        "phpunit/phpunit": "^11.0",
+        "rector/rector": "^2.0"
     },
     "config": {
         "optimize-autoloader": true,
